common

+++ bin/dot-bootstrap.sh
@@ -1,12 +1,12 @@
 #!/usr/bin/env bash
 set -e
 
-if ! [ -x "$(command -v ansible)" ]; then
+if [ -f /etc/redhat-release ] && ! [ -x "$(command -v ansible)" ]; then
   yum install ansible -y
+elif [ -f /etc/debian_version ] && ! [ -x "$(command -v ansible)" ]; then
+  apt update && apt install ansible -y
+elif [ -f /etc/arch-release ] && ! [ -x "$(command -v ansible)" ]; then
+  pacman -Sy ansible --noconfirm
 fi
 
-ansible-playbook -i ~/git/.dotfiles/inventory/hosts.yml ~/git/.dotfiles/plays/deploy.yml --ask-become-pass --diff
-
-#if command -v terminal-notifier 1>/dev/null 2>&1; then
-#  terminal-notifier -title "dotfiles: Bootstrap complete" -message "Successfully set up environment."
-#fi
+ansible-playbook -i ~/git/.dotfiles/inventory/hosts.yml ~/git/.dotfiles/plays/dep.yml --ask-become-pass --diff
